Use an onClick action instead of using a form for logout. Remove the new CSS
