--- src/shared/User/LoginButton.css
@@ -1,13 +1,0 @@
-
-form.logout {
-  display: inline;
-}
-
-div.logout {
-  display: inline;
-}
-
-input.logout {
-  margin: 0;
-  padding: 0;
-}
